fix(u128): TxOutput.amount u64->u128 for 12dp flowers -- full propagation
- TxOutput.amount: u64 -> u128 (premine amounts 1.65e21 overflow u64)
- TxOutputLegacy compat: legacy.amount as u128
- wallet: SpendableUtxo/SendParams/BuildResult/WalletError/select_utxos all u128
- batch: Recipient/BatchResult/MIN_PAYOUT_AMOUNT/select_batch all u128
- burn: circulating_supply_atomic u128, verify_burn_tx->u128, format_zion 12dp
- fee: MAX_OUTPUT_AMOUNT u128=144B*1e12, validate_output_amounts u128
- state: input_sum/output_sum u128, fee as u128 comparison
- validation: total_output u128, Vec<(u128,&str)>
- mempool: Vec<(u128,&str)>
- security_audit: amount as u128, edge_cases Vec<(u128,&str)>
- lmdb: balance_cache (u128,u64), schema v3 w/ ok().flatten() migration, credit_balance u128
- jsonrpc: creditBalance/sendTransaction amount_atomic 1e12 u128, credited_atomic .to_string()
- rpc/methods: balance/premine/vault tx all u128
- zion-wallet: amount_atomic u128, 1e12 conversion, UTXO parse from string

+++ L1/core/src/blockchain/validation.rs
@@ -251,9 +251,9 @@ pub fn validate_block(
         }
 
         // Validate coinbase reward — FEES ARE BURNED, coinbase ≤ block reward only
-        let total_output: u64 = coinbase.outputs.iter().map(|o| o.amount).sum();
+        let total_output: u128 = coinbase.outputs.iter().map(|o| o.amount).sum();
         let max_allowed = fee::max_coinbase_output(block.header.height);
-        if total_output > max_allowed {
+        if total_output > max_allowed as u128 {
             return Err(format!(
                 "Coinbase output {} exceeds max allowed {} (fee burning: fees are NOT paid to miner)",
                 total_output, max_allowed
@@ -311,7 +311,7 @@ pub fn validate_transaction(tx: &crate::tx::Transaction) -> Result<(), String> {
     }
 
     // 4. Output amount validation (no zero, no overflow, within total supply)
-    let outputs: Vec<(u64, &str)> = tx
+    let outputs: Vec<(u128, &str)> = tx
         .outputs
         .iter()
         .map(|o| (o.amount, o.address.as_str()))
